Revert "Change service layout"
This reverts commit 877ea3c2cdb0cdf0ad0971561beb13b61a84160c.
