+++ pkg/services/secrets/database_mig.go
@@ -1,10 +1,9 @@
-package database
+package secrets
 
-import (
-	"github.com/grafana/grafana/pkg/services/sqlstore/migrator"
-)
+import "github.com/grafana/grafana/pkg/services/sqlstore/migrator"
 
-func (ss *SecretsStoreImpl) AddMigration(mg *migrator.Migrator) {
+// AddMigration adds migration
+func (s *SecretsService) AddMigration(mg *migrator.Migrator) {
 	dataKeysV1 := migrator.Table{
 		Name: "data_keys",
 		Columns: []*migrator.Column{
