+++ pkg/services/secrets/secrets_test.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"testing"
 
-	secrets "github.com/grafana/grafana/pkg/services/secrets/manager"
-
 	"github.com/grafana/grafana/pkg/services/secrets/types"
 
 	"github.com/grafana/grafana/pkg/services/secrets/database"
@@ -141,7 +139,7 @@ func TestSecretsService_DataKeys(t *testing.T) {
 	})
 }
 
-func setupSecretService(t *testing.T) secrets.SecretsService {
+func setupSecretService(t *testing.T) SecretsService {
 	t.Helper()
 	raw, err := ini.Load([]byte(`
 [security]
@@ -150,7 +148,7 @@ secret_key = SdlklWklckeLS
 	require.NoError(t, err)
 	settings := &setting.OSSImpl{Cfg: &setting.Cfg{Raw: raw}}
 
-	s := secrets.SecretsService{
+	s := SecretsService{
 		Store:    &database.SecretsStoreImpl{SQLStore: sqlstore.InitTestDB(t)},
 		Enc:      &encryption.OSSEncryptionService{},
 		Settings: settings,
